Explicit check that commits is greater than zero
Krishna managed to hit a case where commits was zero inside DiffPlanCache.
Since we explictly guard for that case when we had previously observed
the plan cache entry, we must have seen a new plan cache entry with zero
successfull commits. This is probably the result of a slow query that has not
yet completed or erroring queries; in either case, we should not assume that
plan cache entries will always have a non-zero commit count.